refactor(entity): merge PageNumber and Version into Ordinal
Both were one value object written twice: a counter with a floor of 1,
differing only in width. Ordinal is that type once, at i32, and page
number and version reach it through the same corrupt-invariant path.

sort_order widens from smallint to integer to match. Migration
20260822120100 changes in place rather than gaining a successor, since
nothing here is merged; a database that already applied it no longer
matches its checksum and needs a rebuild.

order_chapter_pages drops its ::smallint cast. WITH ORDINALITY yields
bigint, which an UPDATE seats in an integer column through an
assignment cast -- the explicit one was never load-bearing.

Wire format is unchanged: version and pageNumber still serialize as
bare numbers.

Refs #10

+++ migrations/20260822120100_add_chapter_pages.up.sql
@@ -1,7 +1,7 @@
 create table if not exists chapter_pages (
 	id uuid,
 	release_id uuid not null,
-	sort_order smallint,
+	sort_order integer,
 	extension text not null,
 
 	constraint pk_chapter_pages_id primary key(id),
